convert all the price math to use Doubles and not Ints

+++ inventory/src/main/java/com/techvvs/inventory/model/CartVO.groovy
@@ -51,7 +51,7 @@ class CartVO implements Serializable{
     String barcode
 
     @JsonProperty
-    Integer total;
+    Double total;
 
     @JsonProperty
     Integer isprocessed;
