+++ inventory/src/main/java/com/techvvs/inventory/model/PaymentVO.groovy
@@ -19,7 +19,7 @@ class PaymentVO implements Serializable {
     Integer paymentid
 
     @JsonProperty
-    Integer amountpaid;
+    Double amountpaid;
 
     @JsonProperty
     String paymenttype;
